+++ Mods ChoGGi/Disable Pause/metadata.lua
@@ -1,0 +1,14 @@
+return PlaceObj("ModDef", {
+	"title", "Disable Pause v0.1",
+	"version", 1,
+	"saved", 1543665600,
+	"id", "ChoGGi_DisablePause",
+	"image", "Preview.png",
+	"steam_id", "1579991229",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[Space sets speed to 1 normal, and the pause button is hidden.]],
+})